requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
psycopg2-binary==2.9.7
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
paypalrestsdk==1.13.3
requests==2.31.0
email-validator==2.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
